move sdk-dapp-ui lib folder
