+++ src/UI/Fields/FeeAccordion/FeeAccordion.tsx
@@ -7,9 +7,9 @@ import useCollapse from 'react-collapsed';
 import globals from 'assets/sass/globals.module.scss';
 import { FormDataTestIdsEnum } from 'constants/formDataTestIds';
 import { useSendFormContext } from 'contexts/SendFormProviderContext';
+import { MvxSpinnerIcon } from 'lib/sdkDappUi';
 import { WithClassnameType } from 'types';
 import { FormatAmount } from 'UI/FormatAmount';
-import { MvxSpinnerIcon } from 'UI/sdkDappUi';
 import { FeeInFiat } from './FeeInFiat';
 import styles from './styles.module.scss';
 import { GasLimit } from '../GasLimit/GasLimit';
